[FIX] stock_barcodes: remove last scan do not revert last action.
In pickings, the scan log was always generated without details of
the move lines modified/created. Without this information in the log
removing the log was dong nothing else than just that, removing the
log; leaving the last operation still reflected in the picking.

Also, when the picking was a draft one, the stock move line and
the stock move could have been created by the scan itself, so we
need to remove it in that case, reverting the action would leave
the stock move there that could be validated by mistake.